Finished TeamCreation.py Cleanup
-Removed traceback imports from several files for consolidation
-Removed redundant imports
-Removed a useless function in TeamCreation.py
-Shifted universal helpers to init.py (should make a new file later for these)
-Other consolidation tactics without affecting logic.
